requirements.txt
# Core dependencies
requests>=2.28.1
beautifulsoup4>=4.11.1
lxml>=4.9.1
PyYAML>=6.0

# For caching and optimization
cachetools>=5.2.0

# For security research
nvdlib>=0.7.4
gitpython>=3.1.30

# For data processing
python-dateutil>=2.8.2

# Optional dependencies for report generation
markdown>=3.4.1
jinja2>=3.1.2

# For command line interface
argparse>=1.4.0

# For semantic version parsing
semver>=2.13.0

smolagents>=0.1.0
ollama>=0.1.7
google-generativeai>=0.3.0 